Add authorization to posts edit and delete methods.

+++ backend/routes/posts.js
@@ -34,8 +34,10 @@ router.post(
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
+
   post.save().then(createdPost => {
     res.status(201).json({
       message: 'Post added successfully',
@@ -64,8 +66,12 @@ router.patch(
       imagePath: imagePath
     });
     console.log(post);
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
-      res.status(200).json({ message: "Update successful!" });
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+      if( result.nModified > 0) {
+        res.status(200).json({ message: "Update successful!" });
+      } else {
+        res.status(401).json({ message: "Update unsuccessful!" });
+      }
     });
   }
 );
@@ -107,9 +113,12 @@ router.delete(
   "/:id",
   checkAuth,
   (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
-    console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+    if( result.n > 0) {
+      res.status(200).json({ message: "Deletion successful!" });
+    } else {
+      res.status(401).json({ message: "Deletion unsuccessful!" });
+    }
   })
 });
